factor out a raw module with all of the unsafe code
