+++ src/core.rs
@@ -1,4 +1,36 @@
-use crate::MAX_LEN;
+use crate::{raw, MAX_1BYTE_TAG, MAX_LEN};
+
+/// Errors that may occur when decoding a `PrefixVarInt`.
+#[derive(Copy, Clone, Debug, Eq, PartialEq)]
+pub enum DecodeError {
+    /// Reached end-of-buffer unexpectedly.
+    ///
+    /// This may happen if you attempt to decode an empty buffer or if the buffer is too short to
+    /// contain the expected value.
+    UnexpectedEob,
+    /// The value read is larger than the destination type.
+    Overflow,
+}
+
+impl std::fmt::Display for DecodeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl std::error::Error for DecodeError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        None
+    }
+
+    fn description(&self) -> &str {
+        ""
+    }
+
+    fn cause(&self) -> Option<&dyn std::error::Error> {
+        None
+    }
+}
 
 /// Base integer trait for prefix varints. Provides internal APIs to facilitate the transforms in
 /// the PrefixVarInt trait.
@@ -18,14 +50,27 @@ impl Int for u64 {
     }
 }
 
+/// Maps negative values to positive values, creating a sequence that alternates between negative
+/// and positive values. This makes the value more amenable to efficient prefix uvarint encoding.
+#[inline]
+pub(crate) fn zigzag_encode(v: i64) -> u64 {
+    ((v >> 63) ^ (v << 1)) as u64
+}
+
+/// Inverts `zigzag_encode()`.
+#[inline]
+pub(crate) fn zigzag_decode(v: u64) -> i64 {
+    (v >> 1) as i64 ^ -(v as i64 & 1)
+}
+
 impl Int for i64 {
     #[inline(always)]
     fn to_prefix_varint_raw(self) -> u64 {
-        crate::zigzag_encode(self)
+        zigzag_encode(self)
     }
     #[inline(always)]
     fn from_prefix_varint_raw(raw: u64) -> Option<Self> {
-        Some(crate::zigzag_decode(raw))
+        Some(zigzag_decode(raw))
     }
 }
 
@@ -49,28 +94,56 @@ impl_int!(u32, u64);
 impl_int!(i16, i64);
 impl_int!(i32, i64);
 
+/// A single encoded prefix varint value for with `PrefixVarInt.to_prefix_varint_bytes()`.
+pub struct EncodedPrefixVarInt {
+    buf: [u8; MAX_LEN],
+    len: u8,
+}
+
+impl EncodedPrefixVarInt {
+    fn new(v: u64) -> Self {
+        let mut enc = Self::default();
+        let len = unsafe { raw::encode(v, enc.buf.as_mut_ptr()) };
+        enc.len = len as u8;
+        enc
+    }
+
+    pub fn as_slice(&self) -> &[u8] {
+        &self.buf[..(self.len as usize)]
+    }
+}
+
+impl Default for EncodedPrefixVarInt {
+    fn default() -> Self {
+        Self {
+            buf: [0u8; MAX_LEN],
+            len: 0,
+        }
+    }
+}
+
 #[inline(always)]
-fn decode_prefix_varint_raw(buf: &[u8]) -> Result<(u64, usize), crate::DecodeError> {
+fn decode_raw(buf: &[u8]) -> Result<(u64, usize), DecodeError> {
     if buf.is_empty() {
-        return Err(crate::DecodeError::UnexpectedEob);
+        return Err(DecodeError::UnexpectedEob);
     }
 
     if buf.len() >= MAX_LEN {
-        return Ok(unsafe { crate::decode_prefix_uvarint(buf.as_ptr()) });
+        return Ok(unsafe { raw::decode(buf.as_ptr()) });
     }
 
     let tag = buf[0];
-    if tag <= crate::MAX_1BYTE_TAG {
+    if tag <= MAX_1BYTE_TAG {
         return Ok((tag.into(), 1));
     }
 
     let len = tag.leading_ones() as usize + 1;
     if len <= buf.len() {
-        let mut ibuf = [0u8; crate::MAX_LEN];
+        let mut ibuf = [0u8; MAX_LEN];
         ibuf[..len].copy_from_slice(&buf[..len]);
-        Ok(unsafe { crate::decode_prefix_uvarint_slow(tag, ibuf.as_ptr()) })
+        Ok(unsafe { raw::decode_multibyte(tag, ibuf.as_ptr()) })
     } else {
-        Err(crate::DecodeError::UnexpectedEob)
+        Err(DecodeError::UnexpectedEob)
     }
 }
 
@@ -80,18 +153,20 @@ fn decode_prefix_varint_raw(buf: &[u8]) -> Result<(u64, usize), crate::DecodeErr
 pub trait PrefixVarInt: Sized + Copy + Int {
     /// Returns the number of bytes required to encode `self`.
     /// This value will always be in `[1, MAX_LEN]`
+    #[inline]
     fn prefix_varint_len(self) -> usize {
-        crate::prefix_uvarint_len(self.to_prefix_varint_raw())
+        raw::len(self.to_prefix_varint_raw())
     }
 
     /// Encode `self` to buf and return the number of bytes written.
     ///
     /// # Panics
     ///
-    /// If `self.prefix_varint_len() > buf.remaining()`.
+    /// If `self.prefix_varint_len() > buf.len()`.
+    #[inline]
     fn encode_prefix_varint(self, buf: &mut [u8]) -> usize {
-        if buf.len() >= crate::MAX_LEN {
-            unsafe { crate::encode_prefix_uvarint(self.to_prefix_varint_raw(), buf.as_mut_ptr()) }
+        if buf.len() >= MAX_LEN {
+            unsafe { raw::encode(self.to_prefix_varint_raw(), buf.as_mut_ptr()) }
         } else {
             let enc = self.to_prefix_varint_bytes();
             let ebytes = enc.as_slice();
@@ -101,18 +176,20 @@ pub trait PrefixVarInt: Sized + Copy + Int {
     }
 
     /// Decode an integer from the bytes in `buf` and return the value and number of bytes consumed.
-    fn decode_prefix_varint(buf: &[u8]) -> Result<(Self, usize), crate::DecodeError> {
-        let (raw, len) = decode_prefix_varint_raw(buf)?;
+    #[inline]
+    fn decode_prefix_varint(buf: &[u8]) -> Result<(Self, usize), DecodeError> {
+        let (raw, len) = decode_raw(buf)?;
         Ok((
-            Self::from_prefix_varint_raw(raw).ok_or(crate::DecodeError::Overflow)?,
+            Self::from_prefix_varint_raw(raw).ok_or(DecodeError::Overflow)?,
             len,
         ))
     }
 
     /// Encode `self` to an owned buffer and return it.
     /// Use `as_slice()` to access the encoded bytes.
-    fn to_prefix_varint_bytes(self) -> crate::EncodedPrefixVarint {
-        crate::EncodedPrefixVarint::new(self.to_prefix_varint_raw())
+    #[inline]
+    fn to_prefix_varint_bytes(self) -> EncodedPrefixVarInt {
+        EncodedPrefixVarInt::new(self.to_prefix_varint_raw())
     }
 }
